requirements.txt
PySimpleGUI-4-foss==4.60.4.1
paramiko==4.0.0
pyserial==3.5
pywin32==311
configparser==7.2.0
matplotlib==3.10.8
numpy==2.4.1
psutil==7.2.1
